Refactor: Move Google Drive functionality to a separate utility file
