+++ src/lib/util/index.ts
@@ -5,3 +5,4 @@ export * from './modals';
 export * from './zip';
 export * from './cloud';
 export * from './sync-store';
+export * from './google-drive';
