+++ src/routes/cloud/+page.svelte
@@ -2,68 +2,42 @@
   import { run } from 'svelte/legacy';
 
   import { processFiles } from '$lib/upload';
-  import { parseVolumesFromJson, profiles, volumes } from '$lib/settings';
+  import { profiles } from '$lib/settings';
   import { miscSettings, updateMiscSetting } from '$lib/settings/misc';
 
-  import { promptConfirmation, showSnackbar, uploadFile, syncFunctionStore } from '$lib/util';
+  import { 
+    promptConfirmation, 
+    showSnackbar, 
+    uploadFile, 
+    accessTokenStore,
+    readerFolderIdStore,
+    volumeDataIdStore,
+    profilesIdStore,
+    tokenClientStore,
+    initGoogleDriveApi,
+    signIn,
+    logout,
+    syncReadProgress
+  } from '$lib/util';
   import { Badge, Button, Toggle } from 'flowbite-svelte';
   import { onMount } from 'svelte';
   import { GoogleSolid } from 'flowbite-svelte-icons';
-  import { progressTrackerStore } from '$lib/util/progress-tracker';
 
-  interface Props {
-    accessToken?: string;
-  }
-
-  let { accessToken = $bindable('') }: Props = $props();
-
-  // Helper function to handle errors consistently
-  function handleDriveError(error: any, context: string) {
-    // Check if it's a connectivity issue
-    const errorMessage = error.toString().toLowerCase();
-    const isConnectivityError =
-      errorMessage.includes('network') ||
-      errorMessage.includes('connection') ||
-      errorMessage.includes('offline') ||
-      errorMessage.includes('internet');
-
-    if (!isConnectivityError) {
-      // Log the user out for non-connectivity errors
-      logout();
-      showSnackbar(`Error ${context}: ${error.message || 'Unknown error'}`);
-    } else {
-      showSnackbar('Connection error: Please check your internet connection');
-    }
-
-    console.error(`${context} error:`, error);
-  }
-
-  const CLIENT_ID = import.meta.env.VITE_GDRIVE_CLIENT_ID;
-  const API_KEY = import.meta.env.VITE_GDRIVE_API_KEY;
-
-  const DISCOVERY_DOC = 'https://www.googleapis.com/discovery/v1/apis/drive/v3/rest';
-  const SCOPES = 'https://www.googleapis.com/auth/drive.file';
-
-  const FOLDER_MIME_TYPE = 'application/vnd.google-apps.folder';
-  const READER_FOLDER = 'mokuro-reader';
-  const VOLUME_DATA_FILE = 'volume-data.json';
-  const PROFILES_FILE = 'profiles.json';
-
-  const type = 'application/json';
-
-  let tokenClient: any;
-  let readerFolderId = '';
+  // Subscribe to stores
+  let accessToken = $state('');
+  let readerFolderId = $state('');
   let volumeDataId = $state('');
   let profilesId = $state('');
+  let tokenClient = $state(null);
+  
+  accessTokenStore.subscribe(value => { accessToken = value; });
+  readerFolderIdStore.subscribe(value => { readerFolderId = value; });
+  volumeDataIdStore.subscribe(value => { volumeDataId = value; });
+  profilesIdStore.subscribe(value => { profilesId = value; });
+  tokenClientStore.subscribe(value => { tokenClient = value; });
 
-  // This variable is used to track if we're connected to Google Drive
-  // and is used in the UI to show/hide the login button
-
-  run(() => {
-    if (accessToken) {
-      localStorage.setItem('gdrive_token', accessToken);
-    }
-  });
+  // Use constants from the google-drive utility
+  const type = 'application/json';
 
   async function getFileSize(fileId: string): Promise<number> {
     try {
@@ -130,137 +104,6 @@
     return xhrDownloadFileIdWithTracking(fileId, fileName, () => {});
   }
 
-  export async function connectDrive(resp?: any) {
-    if (resp?.error !== undefined) {
-      localStorage.removeItem('gdrive_token');
-      accessToken = '';
-      throw resp;
-    }
-
-    accessToken = resp?.access_token;
-
-    const processId = 'connect-drive';
-    progressTrackerStore.addProcess({
-      id: processId,
-      description: 'Connecting to Google Drive',
-      progress: 0,
-      status: 'Initializing connection...'
-    });
-
-    try {
-      progressTrackerStore.updateProcess(processId, {
-        progress: 20,
-        status: 'Checking for reader folder...'
-      });
-
-      const { result: readerFolderRes } = await gapi.client.drive.files.list({
-        q: `mimeType='application/vnd.google-apps.folder' and name='${READER_FOLDER}'`,
-        fields: 'files(id)'
-      });
-
-      if (readerFolderRes.files?.length === 0) {
-        progressTrackerStore.updateProcess(processId, {
-          progress: 40,
-          status: 'Creating reader folder...'
-        });
-
-        const { result: createReaderFolderRes } = await gapi.client.drive.files.create({
-          resource: { mimeType: FOLDER_MIME_TYPE, name: READER_FOLDER },
-          fields: 'id'
-        });
-
-        readerFolderId = createReaderFolderRes.id || '';
-      } else {
-        const id = readerFolderRes.files?.[0]?.id || '';
-        readerFolderId = id || '';
-      }
-
-      progressTrackerStore.updateProcess(processId, {
-        progress: 60,
-        status: 'Checking for volume data...'
-      });
-
-      const { result: volumeDataRes } = await gapi.client.drive.files.list({
-        q: `'${readerFolderId}' in parents and name='${VOLUME_DATA_FILE}'`,
-        fields: 'files(id, name)'
-      });
-
-      if (volumeDataRes.files?.length !== 0) {
-        volumeDataId = volumeDataRes.files?.[0].id || '';
-      }
-
-      progressTrackerStore.updateProcess(processId, {
-        progress: 80,
-        status: 'Checking for profiles...'
-      });
-
-      const { result: profilesRes } = await gapi.client.drive.files.list({
-        q: `'${readerFolderId}' in parents and name='${PROFILES_FILE}'`,
-        fields: 'files(id, name)'
-      });
-
-      if (profilesRes.files?.length !== 0) {
-        profilesId = profilesRes.files?.[0].id || '';
-      }
-
-      progressTrackerStore.updateProcess(processId, {
-        progress: 100,
-        status: 'Connected successfully'
-      });
-      setTimeout(() => progressTrackerStore.removeProcess(processId), 3000);
-
-      if (accessToken) {
-        showSnackbar('Connected to Google Drive');
-        
-        // Check if we need to sync after login
-        const syncAfterLogin = localStorage.getItem('sync_after_login');
-        if (syncAfterLogin === 'true') {
-          // Clear the flag
-          localStorage.removeItem('sync_after_login');
-          // Perform sync
-          setTimeout(() => performSync(), 500);
-        }
-      }
-    } catch (error) {
-      progressTrackerStore.updateProcess(processId, {
-        progress: 0,
-        status: 'Connection failed'
-      });
-      setTimeout(() => progressTrackerStore.removeProcess(processId), 3000);
-      handleDriveError(error, 'connecting to Google Drive');
-    }
-  }
-
-  function signIn() {
-    // Always show the account picker to allow switching accounts
-    tokenClient.requestAccessToken({ prompt: 'consent' });
-  }
-
-  export function logout() {
-    // Remove token from localStorage
-    localStorage.removeItem('gdrive_token');
-
-    // Clear the token from memory
-    accessToken = '';
-
-    // Revoke the token with Google to ensure account picker shows up next time
-    if (gapi.client.getToken()) {
-      const token = gapi.client.getToken().access_token;
-      // Clear the token from gapi client
-      gapi.client.setToken(null);
-
-      // Revoke the token with Google's OAuth service
-      fetch(`https://oauth2.googleapis.com/revoke?token=${token}`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/x-www-form-urlencoded'
-        }
-      }).catch((error) => {
-        console.error('Error revoking token:', error);
-      });
-    }
-  }
-
   // Function to clear service worker cache for Google Drive downloads
   async function clearServiceWorkerCache() {
     if ('caches' in window) {
@@ -301,45 +144,13 @@
   }
 
   onMount(() => {
-    // Register the sync function in the store
-    syncFunctionStore.set(syncVolumeData);
-    
     // Clear service worker cache for Google Drive downloads
     clearServiceWorkerCache();
     
-    gapi.load('client', async () => {
-      try {
-        await gapi.client.init({
-          apiKey: API_KEY,
-          discoveryDocs: [DISCOVERY_DOC]
-        });
-
-        // Initialize token client after gapi client is ready
-        tokenClient = google.accounts.oauth2.initTokenClient({
-          client_id: CLIENT_ID,
-          scope: SCOPES,
-          callback: connectDrive
-        });
-
-        // Try to restore the saved token only after gapi client is initialized
-        const savedToken = localStorage.getItem('gdrive_token');
-        if (savedToken) {
-          try {
-            // Set the token in gapi client
-            gapi.client.setToken({ access_token: savedToken });
-            accessToken = savedToken;
-            await connectDrive({ access_token: savedToken });
-          } catch (error) {
-            console.error('Failed to restore saved token:', error);
-            // Token will be cleared in connectDrive if there's an error
-          }
-        }
-      } catch (error) {
-        handleDriveError(error, 'initializing Google Drive');
-      }
+    // Initialize Google Drive API
+    initGoogleDriveApi().catch(error => {
+      console.error('Failed to initialize Google Drive API:', error);
     });
-
-    gapi.load('picker', () => {});
   });
 
   function createPicker() {
@@ -816,139 +627,9 @@
     }
   }
   
-  // This function handles the sync operation with authentication check
-  export async function syncVolumeData() {
-    // If not authenticated, prompt login first
-    if (!accessToken) {
-      // Store that we want to sync after login
-      localStorage.setItem('sync_after_login', 'true');
-      signIn();
-      return;
-    }
-    
-    await performSync();
-  }
-  
-  // This function performs the actual sync operation
-  async function performSync() {
-    const processId = 'sync-volume-data';
-    progressTrackerStore.addProcess({
-      id: processId,
-      description: 'Syncing read progress',
-      progress: 0,
-      status: 'Starting sync...'
-    });
-
-    try {
-      // Step 1: Download cloud volume data if it exists
-      let cloudVolumes = {};
-      if (volumeDataId) {
-        progressTrackerStore.updateProcess(processId, {
-          progress: 20,
-          status: 'Downloading cloud data...'
-        });
-
-        try {
-          const { body } = await gapi.client.drive.files.get({
-            fileId: volumeDataId,
-            alt: 'media'
-          });
-          cloudVolumes = parseVolumesFromJson(body);
-        } catch (error) {
-          console.error('Error downloading cloud volume data:', error);
-          // Continue with empty cloud volumes if download fails
-        }
-      }
-
-      // Step 2: Get local volume data
-      progressTrackerStore.updateProcess(processId, {
-        progress: 40,
-        status: 'Merging data...'
-      });
-
-      let localVolumes = {};
-      volumes.subscribe(value => {
-        localVolumes = { ...value };
-      })();
-
-      // Step 3: Merge the data, keeping the newer records based on lastProgressUpdate
-      const mergedVolumes = {};
-      
-      // Process all volume IDs from both sources
-      const allVolumeIds = new Set([
-        ...Object.keys(localVolumes),
-        ...Object.keys(cloudVolumes)
-      ]);
-      
-      allVolumeIds.forEach(volumeId => {
-        const localVolume = localVolumes[volumeId];
-        const cloudVolume = cloudVolumes[volumeId];
-        
-        // If volume exists only in one source, use that one
-        if (!localVolume) {
-          mergedVolumes[volumeId] = cloudVolume;
-        } else if (!cloudVolume) {
-          mergedVolumes[volumeId] = localVolume;
-        } else {
-          // Both exist, compare lastProgressUpdate dates
-          const localDate = new Date(localVolume.lastProgressUpdate).getTime();
-          const cloudDate = new Date(cloudVolume.lastProgressUpdate).getTime();
-          
-          // Keep the newer record
-          mergedVolumes[volumeId] = localDate >= cloudDate ? localVolume : cloudVolume;
-        }
-      });
-      
-      // Step 4: Update local volume data with merged data
-      progressTrackerStore.updateProcess(processId, {
-        progress: 60,
-        status: 'Updating local data...'
-      });
-      
-      volumes.update(() => mergedVolumes);
-      
-      // Step 5: Upload merged data to cloud
-      progressTrackerStore.updateProcess(processId, {
-        progress: 80,
-        status: 'Uploading merged data...'
-      });
-      
-      const metadata = {
-        mimeType: type,
-        name: VOLUME_DATA_FILE,
-        parents: [volumeDataId ? null : readerFolderId]
-      };
-      
-      const res = await uploadFile({
-        accessToken,
-        fileId: volumeDataId,
-        metadata,
-        localStorageId: 'volumes',
-        type
-      });
-      
-      volumeDataId = res.id;
-      
-      progressTrackerStore.updateProcess(processId, {
-        progress: 100,
-        status: 'Sync complete'
-      });
-      setTimeout(() => progressTrackerStore.removeProcess(processId), 3000);
-      
-      showSnackbar('Read progress synced successfully');
-    } catch (error) {
-      progressTrackerStore.updateProcess(processId, {
-        progress: 0,
-        status: 'Sync failed'
-      });
-      setTimeout(() => progressTrackerStore.removeProcess(processId), 3000);
-      handleDriveError(error, 'syncing read progress');
-    }
-  }
-  
   // For backward compatibility with the button in the cloud page
-  async function onSyncVolumeData() {
-    await performSync();
+  async function performSync() {
+    await syncReadProgress();
   }
 
   async function onDownloadProfiles() {
